==> notebooks/2_create_study_mission_nodes.ipynb <==
 🌐 Create Study & Mission Nodes in the SPOKE‑GeneLab Knowledge Graph

This notebook reads the GeneLab dataset manifest file, extracts mission‑ and study‑level metadata, and writes Neo4j node and relationship files for integration into the SPOKE‑GeneLab Knowledge Graph via the `genelab_utils` package.

Author: Peter W. Rose, UC San Diego (pwrose.ucsd@gmail.com)

import pandas as pd
import genelab_utils as gl

pd.set_option('display.max_rows', None)  # Shows all rows
pd.set_option('display.max_colwidth', None)  # Shows full content of each cell

## Setup Environment Variables
Edit `../.env` to configure the environment.    

# Node and relationship directory paths
node_dir, rel_dir = gl.setup_environment()

## Validate the KG Metadata Files in the `../kg/v#.#.#/metadata` Directory

gl.validate_kg_metadata()

## Get Info about available Datasets

MANIFEST_PATH = "../data/manifest.csv" # file with dataset info

manifest = pd.read_csv(MANIFEST_PATH)
manifest.head()

## Get GeneLab Mission and Study Metadata

metadata = gl.get_metadata(manifest)
metadata.head()

## Create Mission Nodes

missions = metadata[["mission_id", "name", "flight_program", "space_program", "start_date", "end_date"]]
missions = missions[missions["name"] != ""].copy()
missions.rename(columns={"mission_id": "identifier"}, inplace=True)

mission_nodes = gl.save_dataframe_to_kg(missions, 'Mission', node_dir)
print(f"Number of Mission nodes: {mission_nodes.shape[0]}")
mission_nodes.head()

## Create Study Nodes

studies = metadata[["identifier", "project_title", "project_type", "organism", "taxonomy"]].copy()
studies["name"] = studies["identifier"]
studies = studies[["identifier", "name", "project_title", "project_type", "organism", "taxonomy"]]

study_nodes = gl.save_dataframe_to_kg(studies, 'Study', node_dir)
print(f"Number of Study nodes: {study_nodes.shape[0]}")
study_nodes.head()

## Create Missions-CONDUCTED_MIcS-Study Relationships

mission_conducted_study = metadata[["mission_id", "identifier"]]
# Not all studies have an associated mission (e.g., ground studies)
mission_conducted_study = mission_conducted_study[mission_conducted_study["mission_id"] != ""].copy()
mission_conducted_study.rename(columns={"mission_id": "from", "identifier": "to", }, inplace=True)

mission_conducted_study_rels = gl.save_dataframe_to_kg(mission_conducted_study, 'Mission-CONDUCTED_MIcS-Study', rel_dir)
print(f"Number of Mission-CONDUCTED_MIcS-Study relationships: {mission_conducted_study_rels.shape[0]}")
mission_conducted_study_rels.head()
