==> notebooks/3_create_gene_nodes.ipynb <==
# 🌱 Create Model‑Organism & Human Gene Nodes with Orthology Links

This notebook processes the GeneLab datasets to build Neo4j node and relationship files for model‑organism genes (MGene), human ortholog genes, and their orthology links. It uses the `genelab_utils` and `ortholog_mapper` packages to extract gene IDs, map to human orthologs, and write CSVs ready for SPOKE ingestion.

Author: Peter W. Rose, UC San Diego (pwrose.ucsd@gmail.com)

import pandas as pd
import genelab_utils as gl
import ortholog_mapper

pd.set_option('display.max_rows', None)  # Shows all rows
pd.set_option('display.max_colwidth', None)  # Shows full content of each cell

## Setup Environment Variables
Edit `../.env` to configure the environment.   

# Node and relationship directory paths
node_dir, rel_dir = gl.setup_environment()

MANIFEST_PATH = "../data/manifest.csv" # file with dataset info

## Get Info about available Datasets

manifest = pd.read_csv(MANIFEST_PATH)
manifest.head()

## Create MGene (Model Organism Gene) Nodes

# Parse all data files for gene ids (ENTREZID) and create a unique list of all genes.
mgenes = gl.extract_gene_info(manifest)

mgene_nodes = gl.save_dataframe_to_kg(mgenes, 'MGene', node_dir)
print(f"Number of MGene nodes: {mgene_nodes.shape[0]}")
mgene_nodes.head()

## Map Model Organism Genes to Human Orthologs
Orthologs are mapped using the taxonomy id and ENTREZ gene identifier. Suggestions for suitable ortholog databases are listed below.

# List of taxonomies in the dataset with statistically significant data that have ENTREZ IDs
print(f"List of taxonomy ids in the datasets: {mgenes['taxonomy'].unique()}")

# List of supported ortholog_dbs that have a mapping to human ENTREZ IDs
suggestions = ortholog_mapper.suggest_ortholog_dbs(mgenes, "taxonomy", "identifier")
suggestions

mapped_genes = ortholog_mapper.map_orthologs(mgenes, "taxonomy", "identifier", "human_entrez_id", ortholog_dbs=["JAX", "Ensembl"])
# Remove any genes that cannot be mapped to human ENTREZ ids
mapped_genes = mapped_genes[mapped_genes["human_entrez_id"] != ""]

## Create Human Ortholog Gene Nodes
The mapped genes include non-protein coding genes

human_genes = mapped_genes[['human_entrez_id']].copy()
human_genes.rename(columns={'human_entrez_id': 'identifier'}, inplace=True)

human_gene_nodes = gl.save_dataframe_to_kg(human_genes, 'Gene', node_dir)
print(f"Number of Gene nodes: {human_gene_nodes.shape[0]}")
human_gene_nodes.head()

## Create MGene-IS_ORTHOLOG_MGiG-Gene Relationships

model_to_human_genes = mapped_genes[["identifier", "human_entrez_id"]].copy()
model_to_human_genes.rename(columns={"identifier": "from", "human_entrez_id": "to"}, inplace=True)

model_to_human_gene_rels = gl.save_dataframe_to_kg(model_to_human_genes, 'MGene-IS_ORTHOLOG_MGiG-Gene', rel_dir)
print(f"Number of MGene-IS_ORTHOLOG_MGiG-Gene relationships: {model_to_human_gene_rels.shape[0]}")
model_to_human_gene_rels.head()
