==> notebooks/5_import_to_neo4j.ipynb <==
# 🌐 Import & Inspect SPOKE‑GeneLab Knowledge Graph in Neo4j

This notebook bulk‑imports the CSV node and relationship exports from the SPOKE‑GeneLab pipeline into a Neo4j graph database, then runs exploratory queries to validate schema, metadata, counts, and full‑text search functionality. It supports downloading into a Neo4j Desktop database or Neo4j Enterprise instance (see cell [3])

**Start the `spoke-genelab` Graph DBMS before you run this notebook!**

Author: Peter W. Rose, UC San Diego (pwrose.ucsd@gmail.com)

import os
import pandas as pd
from py2neo import Graph
import neo4j_utils
import neo4j_bulk_importer

pd.set_option('display.max_rows', None)  # Shows all rows
pd.set_option('display.max_columns', None) # Shows all columns
pd.set_option('display.max_colwidth', None)  # Shows full content of each cell

### Import the Knowledge Graph
CSV data and metadata files are uploaded into the Neo4j Graph database from the [kg](https://github.com/BaranziniLab/spoke_genelab/tree/main/kg) directory using the [kg-import](https://github.com/sbl-sdsc/kg-import) bulk upload scripts. For a description of the data organization and the specification of metadata [see](https://github.com/sbl-sdsc/kg-import/blob/main/README.md).

neo4j_bulk_importer.import_from_csv_to_neo4j_desktop(verbose=True) # Use for Neo4j Desktop
# neo4j_bulk_importer.import_from_csv_to_neo4j_enterprise(verbose=True) # Use for Neo4j Enterprise instance

### Connect to the local Neo4j Graph database

database = os.environ.get("NEO4J_DATABASE")
username = os.environ.get("NEO4J_USERNAME")
password = os.environ.get("NEO4J_PASSWORD")
stylesheet = os.environ.get("NEO4J_STYLESHEET")

graph = Graph("bolt://localhost:7687", name=database, user=username, password=password)

## Metadata <a class="anchor" id="Metadata"></a>

### Node metadata
spoke_genelab KG is a self-describing KG. The MetaNodes and MetaRelationships define the structure of the KG and the properties of nodes and relationships. The query below lists the nodes and their properties.

query = """
MATCH (n:MetaNode) RETURN n;
"""
df = graph.run(query).to_data_frame()
metadata = df["n"].tolist()
metadata = pd.DataFrame(metadata)
metadata.fillna("", inplace=True)
metadata

### Metagraph <a class="anchor" id="Metagraph"></a>
The metagraph shows the node labels and relationship types of the KG. Click on a node to display the node metadata.

query = """
MATCH p=(:MetaNode)-->(:MetaNode) RETURN p
"""
subgraph1 = graph.run(query).to_subgraph()

# If the layout isn’t ideal, rerun this cell.
# Use the mouse to drag nodes, pan or zoom the canvas, and adjust positions.
# Click a node to view its properties.
widget1 = neo4j_utils.draw_graph(subgraph1, stylesheet)
widget1.layout.height = "1014px"
widget1.set_layout(name="cola", padding=00, nodeSpacing=65, nodeDimensionsIncludeLabels=True, unconstrIter=15000)
display(widget1)

### Number of Nodes

query = """
MATCH (n) RETURN COUNT(n);
"""
print(f"Total number of nodes: {graph.evaluate(query)}")

query = """
MATCH (n) RETURN labels(n)[0] AS Node, COUNT(n) AS Count
ORDER BY Count DESC;
"""
graph.run(query).to_data_frame()

### Number of relationships

query = """
MATCH (n1)-[r]->(n2)
WITH TYPE(r)        AS Relationship,
     count(r)       AS Count,
     LABELS(n1)[0]  AS Node1,
     LABELS(n2)[0]  AS Node2
RETURN Node1, Relationship, Node2, Count
ORDER BY Count DESC;
"""
graph.run(query).to_data_frame()

## Fulltext Search <a class="anchor" id="Fulltext Search"></a>

### Full text query by keyword or phrase
A full text query returns all nodes that match the text query ([Query Syntax](https://lucene.apache.org/core/5_5_0/queryparser/org/apache/lucene/queryparser/classic/package-summary.html#Overview)). For exact matches, enclose the phrase in double quotes, e.g., ```'"eye"'```.

[Learn more about full-text searches.](https://graphaware.com/neo4j/2019/01/11/neo4j-full-text-search-deep-dive.html)

phrase = '"eye"'
top_n = 5 # show top n hits

query = """
CALL db.index.fulltext.queryNodes("fulltext", $phrase) YIELD node, score
RETURN node.identifier AS identifier, LABELS(node)[0] AS type, node;
"""
graph.run(query, phrase=phrase).to_data_frame().head(top_n)

### Full text query using boolean operators
The full text query supports a variety of query types, including fuzzy, proximity, and range queries, as well as boolean operators ([Query Syntax](https://lucene.apache.org/core/5_5_0/queryparser/org/apache/lucene/queryparser/classic/package-summary.html#Overview)). The following example uses a query with an ```AND``` operator.

phrase = 'liver AND "DNA methylation profiling"'

query = """
CALL db.index.fulltext.queryNodes("fulltext", $phrase) YIELD node, score
RETURN node.identifier AS identifier, LABELS(node)[0] AS type, node, score
ORDER BY type;
"""
graph.run(query, phrase=phrase).to_data_frame()
